+++ doc/Changelog.md
@@ -1,4 +1,5 @@
 * added navigation in albums with enter and left/arrow arrow keys
+* identify root album in search result
 
 ### version v5.0.0 (Apr 27, 2021)